+++ CHANGELOG.md
@@ -17,4 +17,4 @@ File changes:
 - `anpwrapper.h`:
 - `tcp.c`:
 - `tcp.h`:
-
+- 'ip_rx.c': Implemented logic for tcp_rx